Delete
Per direction of M365DEV, the learning path & modules deleted in this PR were not supposed to be published as is. All four modules were supposed to be a single module for "Office add-in fundamental" and added to the "m365 fundamental" learning path.

The new module & learning paths have been created & published. This PR cleans up the old incorrect modules.