Update 3-describe-teams-administration.md
Added a note about the addition Y prompts. One is for nugget proivder, the other is to trust the PSGallery Repo.

Also removed the old powershell commands

The New-CSOnlineSession power shell commands are not valid for the teams module.

Tested It with just the Microsoft-Teams Module installed.

Formated the addition Commands to match case. Same and not an issue for matches the format of the module.