Update 2-explore-audio-video-quality-issues.md
Never heard of a packet snipper... packet sniffers like wireshark or IDS or IPS are real and everywhere. I also disagree with this tip in general. Would like to discuss with anyone on why it is in here.  Any device that introduce packet loss should be avoided but most modern day firewalls use ALG devices.  A packet sniffer if pass and didn't introduce packet loss would be no issue at all.